@@ -13,8 +13,8 @@
   "dependencies": {
     "d3-ease": "^1.0.5",
     "react": "^16.8.6",
-    "react-app-polyfill": "^0.2.2",
-    "react-dev-utils": "^8.0.0",
+    "react-app-polyfill": "^1.0.0",
+    "react-dev-utils": "^9.0.0",
     "react-dom": "^16.8.6",
     "react-spring": "^8.0.19",
     "three": "^0.104.0",
@@ -22,8 +22,8 @@
     "vec-la": "^1.5.0"
   },
   "devDependencies": {
-    "@babel/core": "7.4.3",
-    "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/core": "7.4.4",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.3.3",
@@ -34,11 +34,11 @@
     "image-webpack-loader": "^4.6.0",
     "style-loader": "^0.23.1",
     "terser-webpack-plugin": "1.2.3",
-    "webpack": "4.29.6",
-    "webpack-bundle-analyzer": "^3.1.0",
-    "webpack-cli": "^3.3.0",
+    "webpack": "4.30.0",
+    "webpack-bundle-analyzer": "^3.3.2",
+    "webpack-cli": "^3.3.1",
     "webpack-common-shake": "^2.1.0",
-    "webpack-dev-server": "3.2.1"
+    "webpack-dev-server": "3.3.1"
   },
   "scripts": {
     "start": "webpack-dev-server --progress --colors --env development",
